requirements.txt
fastapi
uvicorn
python-multipart
jinja2
python-dotenv
requests
pymongo
qdrant-client
sqlalchemy
sentence-transformers
langchain-text-splitters
google-generativeai
sherpa-onnx
soundfile
numpy
moviepy
beautifulsoup4
browser-cookie3
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
grpcio==1.68.1
grpcio-tools==1.68.1
bcrypt

# OCR Dependencies
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
scikit-image>=0.21.0